backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-dotenv==0.19.0
supabase==1.0.3
notion-client==1.0.0
google-api-python-client==2.31.0
pydantic==1.8.2
python-multipart==0.0.5
